(SIMP-1167) Relocation of 'site_files'
This adapts to the new 'site_files' directory and also takes care of a
bug in the SELinux policy included with EL7.

Addtionally, the SELinux contexts were fixed and the ordering was
straightened out between the KDC service and the configuration items.

SIMP-1167 #comment Several fixes
SIMP-1376 #close
SIMP-1402 #close
SIMP-1412 #close

+++ lib/puppet/provider/krb5kdc_auto_keytabs/generate.rb
@@ -189,7 +189,6 @@ Puppet::Type.type(:krb5kdc_auto_keytabs).provide :generate do
 
       if princ[:create]
         cmd = %(#{command(:kadmin)} -q "add_principal -randkey +allow_renewable +allow_svr #{princ_name}")
-
         Puppet.debug("Running: #{cmd}")
         execute(cmd)
       end
@@ -212,14 +211,20 @@ Puppet::Type.type(:krb5kdc_auto_keytabs).provide :generate do
             fh.puts(vno)
           end
         end
-
-        FileUtils.mv(princ[:tmp_keytab], princ[:keytab])
-        FileUtils.chmod(0640, princ[:keytab])
-        FileUtils.chown(user, group, princ[:keytab])
-        FileUtils.chown(user, group, princ[:vno_file])
       else
         Puppet.warn("Could not add '#{princ_name}' to keytab at '#{princ[:keytab]}'")
         FileUtils.rm_f(princ[:tmp_keytab])
+      end
+    end
+
+    # This is some ugly foo to ensure the tmp .krb5.keytab file is written to
+    # krb5.keytab once.
+    @principals_to_process.keys.sort.each do |princ_name|
+      princ = @principals_to_process[princ_name]
+      FileUtils.mv(princ[:tmp_keytab], princ[:keytab]) if File.exist?(princ[:tmp_keytab])
+      [princ[:keytab],princ[:vno_file]].each do |f|
+        FileUtils.chmod(0640,f)
+        FileUtils.chown(user,group,f)
       end
     end
 
@@ -280,24 +285,54 @@ Puppet::Type.type(:krb5kdc_auto_keytabs).provide :generate do
 
     search_paths = [
       'keydist',
-      File.join('site_files','pki_files','files','keydist')
+      File.join(
+        '/var',
+        'simp',
+        'environments',
+        Puppet[:environment],
+        'site_files',
+        'pki_files',
+        'files',
+        'keydist'
+      )
     ]
 
-    base_dir = Puppet[:confdir]
+    base_dirs = [Puppet[:confdir]]
 
     if Puppet[:environmentpath] && Puppet[:environment]
-      env_path = File.join(Puppet[:environmentpath], Puppet[:environment])
+      env_paths = Puppet[:environmentpath].split(':')
+      env_paths.map!{|x| File.join(x, Puppet[:environment])}
 
-      if File.directory?(env_path)
-        base_dir = env_path
+      unless env_paths.empty?
+        base_dirs = []
+
+        env_paths.each do |env_path|
+          if File.directory?(env_path)
+            base_dirs << env_path
+          end
+        end
       end
     end
 
-    return discovered_hosts unless File.directory?(base_dir)
+    found_base_dir = false
+    base_dirs.each do |base_dir|
+      if File.directory?(base_dir)
+        found_base_dir = true
+        break
+      end
+    end
+
+    return discovered_hosts unless found_base_dir
+
+    search_paths = search_paths.map do |path|
+      if path[0].chr == '/'
+        path = path
+      else
+        path = base_dirs.map{|x| File.join(x, path)}
+      end
+    end.flatten.compact
 
     search_paths.each do |path|
-      path = File.join(base_dir,path)
-
       if File.directory?(path)
         discovered_hosts += Dir.glob(File.join(path, '*.*')).map{|x| File.basename(x)}
       end
